Changed the column name of the ID field.
Set the size of the window for the displaying passwords.

Made the errorMessage method static.
Added edit button and menu item.
Made the ID field size static and small.
Added an inner compound border on the Profiles JList to create inner padding for the text.
Updated the showInputWindow and created the showMessageWindow.